[TWINFACES-288] feat: remove pages from worspace to screens
