+++ src/screens/factory/index.ts
@@ -1,0 +1,1 @@
+export { FactoryScreen } from "./factory";
