+++ src/screens/factory/views/factory-general.tsx
