+++ src/screens/factory/views/index.ts
